Refactor call audio route UI helpers
