+++ client-ios/Resources/en.lproj/Localizable.strings
@@ -185,14 +185,14 @@
 "frontline_stop_screen_share" = "Stop screen share";
 "frontline_share_screen" = "Share screen";
 "frontline_close" = "Close";
-"frontline_audio_route" = "Audio";
-"frontline_audio_speaker" = "Speaker";
-"frontline_audio_earpiece" = "Phone";
-"frontline_audio_headset" = "Headset";
-"frontline_audio_bluetooth" = "Bluetooth";
-"frontline_audio_car" = "Car audio";
-"frontline_audio_usb" = "USB audio";
-"frontline_audio_unknown" = "Audio";
+"call_audio_route" = "Audio";
+"call_audio_speaker" = "Speaker";
+"call_audio_phone" = "Phone";
+"call_audio_headset" = "Headset";
+"call_audio_bluetooth" = "Bluetooth";
+"call_audio_car" = "Car audio";
+"call_audio_usb" = "USB audio";
+"call_audio_unknown" = "Audio";
 
 "feedback_title" = "Feedback";
 "feedback_send_action" = "Send Feedback";
