+++ client-ios/Resources/es.lproj/Localizable.strings
@@ -185,14 +185,14 @@
 "frontline_stop_screen_share" = "Detener pantalla compartida";
 "frontline_share_screen" = "Compartir pantalla";
 "frontline_close" = "Cerrar";
-"frontline_audio_route" = "Audio";
-"frontline_audio_speaker" = "Altavoz";
-"frontline_audio_earpiece" = "Teléfono";
-"frontline_audio_headset" = "Auriculares";
-"frontline_audio_bluetooth" = "Bluetooth";
-"frontline_audio_car" = "Audio del coche";
-"frontline_audio_usb" = "Audio USB";
-"frontline_audio_unknown" = "Audio";
+"call_audio_route" = "Audio";
+"call_audio_speaker" = "Altavoz";
+"call_audio_phone" = "Teléfono";
+"call_audio_headset" = "Auriculares";
+"call_audio_bluetooth" = "Bluetooth";
+"call_audio_car" = "Audio del coche";
+"call_audio_usb" = "Audio USB";
+"call_audio_unknown" = "Audio";
 
 "feedback_title" = "Comentarios";
 "feedback_send_action" = "Enviar comentarios";
