+++ client-ios/Resources/ru.lproj/Localizable.strings
@@ -185,14 +185,14 @@
 "frontline_stop_screen_share" = "Остановить демонстрацию экрана";
 "frontline_share_screen" = "Показать экран";
 "frontline_close" = "Закрыть";
-"frontline_audio_route" = "Аудио";
-"frontline_audio_speaker" = "Динамик";
-"frontline_audio_earpiece" = "Телефон";
-"frontline_audio_headset" = "Гарнитура";
-"frontline_audio_bluetooth" = "Bluetooth";
-"frontline_audio_car" = "Автоаудио";
-"frontline_audio_usb" = "USB-аудио";
-"frontline_audio_unknown" = "Аудио";
+"call_audio_route" = "Аудио";
+"call_audio_speaker" = "Динамик";
+"call_audio_phone" = "Телефон";
+"call_audio_headset" = "Гарнитура";
+"call_audio_bluetooth" = "Bluetooth";
+"call_audio_car" = "Автоаудио";
+"call_audio_usb" = "USB-аудио";
+"call_audio_unknown" = "Аудио";
 
 "feedback_title" = "Отзыв";
 "feedback_send_action" = "Отправить отзыв";
